Fix allocation header bug found in TArray test
- Fix allocation header handling of alignment (allocation bug in TArray ranged loop test).
- Move allocation header functions to HeapAllocator.
- Move DefAlloc definition to IAlloc.cpp.
- Remove MemoryInternal.h and MemoryInternal.cpp
- Change const -> constexpr in HeapAllocatorTests.cpp.
