+++ Core/Core.vcxproj
@@ -24,14 +24,12 @@
     <ClInclude Include="Memory\Containers\TArray.h" />
     <ClInclude Include="Memory\IAlloc.h" />
     <ClInclude Include="Memory\HeapAllocator.h" />
-    <ClInclude Include="Memory\MemoryInternal.h" />
     <ClInclude Include="Utility.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Memory\AllocTracker.cpp" />
     <ClCompile Include="Memory\HeapAllocator.cpp" />
     <ClCompile Include="Memory\IAlloc.cpp" />
-    <ClCompile Include="Memory\MemoryInternal.cpp" />
     <ClCompile Include="Utility.cpp" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
